Use DATETIME function when querying feeds
Yet another update in an attempt to fix problems with the query to find
feeds that need checking, (in feed_utils.find_feeds_to_check().

This change works right now but over the course of a few days it seems to
regress. For some reason the day used in the query is yesterday rather than
today, i.e. the day in working_date. Need to keep an eye on this.