Ensure working_date is now
By setting working_date to utcnow() in the function declaration it was
being set at the time the function was declared, i.e. what the app is
started by GAE. Since the app isn't restarted very often it means this
time is usually wrong.

With this change I'm defaulting working_date to None in the function
declaration and the updating it to utcnow() in the function body.

See https://stackoverflow.com/a/33938617/219525